requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.7.0
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
psycopg==3.2.6
pydot==3.0.4
PyJWT==2.9.0
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
